Parallel online Rhat (#17)
This PR implements a fully parallel online Rhat for independent parallel MCMC inference, enabling brute force cross validation on an interactive timescale.

In its simpler form (without stabilization with a sigmoid transform) it matches Arviz's split Rhat implementation, but uses O(1) memory. That is, the MCMC draws are not retained on the GPU. However, the implementation has been careful with random number generator state, so problematic posteriors can be instantly replayed on CPU for diagnostic analysis.

The PR also removes a bunch of code that didn't turn out to be useful. This version relies on the GHMC and MEADS implementations of Blackjax.